knapsack: make per-call the default, add KnapsackUpfront opt-in (#200)
Swap which implementation is canonical, on measured evidence: the
per-call DP proofs verify 3.6-18x faster than the upfront-DAG reimpl,
so make per-call the default `Knapsack` and keep the upfront reimpl as
the opt-in `KnapsackUpfront` (3-6x smaller proofs, for size-critical
use).

Two-way rename swap (both classes, tests, and cake-conformance kept):
  KnapsackLegacy -> Knapsack          (per-call, now the default)
    knapsack_legacy.{hh,cc}      -> knapsack.{hh,cc}
    knapsack_legacy_test.cc      -> knapsack_test.cc
    constraint_type/s_expr/hint  -> "knapsack"
  Knapsack       -> KnapsackUpfront    (upfront-DAG, now the opt-in)
    knapsack.{hh,cc}             -> knapsack_upfront.{hh,cc}
    knapsack_test.cc             -> knapsack_upfront_test.cc
    constraint_type/s_expr/hint  -> "knapsack_upfront"

Default construction sites (xcsp, minizinc, examples/knapsack, the
primary knapsack_test, scp_reader's `knapsack` keyword) already post
`Knapsack`, which now resolves to the per-call default; the upfront
variant is exercised by knapsack_upfront_test and `knapsack_bench
--upfront` (flag renamed from --legacy). scp_reader's `knapsack`
keyword and the scp_cases knapsack_sat/unsat chain tests now exercise
the per-call default and still cake-chain-verify. dev_docs/knapsack.md
reframed around the measured default-vs-opt-in trade-off.
