+++ gcs/constraints/knapsack.hh
@@ -2,6 +2,6 @@
 #define GLASGOW_CONSTRAINT_SOLVER_GUARD_GCS_CONSTRAINTS_KNAPSACK_HH
 
 #include <gcs/constraints/knapsack/knapsack.hh>
-#include <gcs/constraints/knapsack/knapsack_legacy.hh>
+#include <gcs/constraints/knapsack/knapsack_upfront.hh>
 
 #endif
